Giant mess but it works, mostly
Need to redo the ORM sequlize file structure to be more in line with ECMA6. some of it is some of it currently is not. Kind of a mess but I wanted a save point before things get real messy. This is all being done so we can get custom validation messages returning from squizel in the correct language.